**2020.2.25-dev (202002250)** - Added alternate `curl` setup instructions to README.md - Default switch_delay: 3.5 seconds - Fixed typos - Updated module framework > Note: incompatible with AccA versions lower than 1.0.21

+++ build.sh
@@ -7,8 +7,6 @@
 #   e.g.,
 #     build.sh (builds $id and generates installable archives)
 #     build.sh any_random_arg (only builds $id)
-#
-# devs: triple hashtags (###) mark custom code
 
 
 (cd ${0%/*} 2>/dev/null
@@ -25,52 +23,53 @@ set_prop() {
 id=$(sed -n "s/^id=//p" module.prop)
 
 version=$(grep '\*\*.*\(.*\)\*\*' README.md \
-  | head -n1 | sed 's/\*\*//; s/ .*//')
+  | head -n 1 | sed 's/\*\*//; s/ .*//')
 
 versionCode=$(grep '\*\*.*\(.*\)\*\*' README.md \
-  | head -n1 | sed 's/\*\*//g; s/.* //' | tr -d ')' | tr -d '(')
+  | head -n 1 | sed 's/\*\*//g; s/.* //' | tr -d ')' | tr -d '(')
+
+tmpDir=.tmp/META-INF/com/google/android
 
 
-set_prop version $version
-set_prop versionCode $versionCode
+# update module.prop
+if ! grep -q "$versionCode" module.prop; then
+  set_prop version $version
+  set_prop versionCode $versionCode
+fi
 
 
 # set ID
-for file in ./install-*.sh ./$id/*.sh ./bundle.sh \
-  ./uninstaller-src/META-INF/com/google/android/update-binary
-do
+for file in ./install-*.sh ./$id/*.sh ./bundle.sh; do
   if [ -f "$file" ] && grep -Eq '(^|\()id=' $file; then
     grep -Eq "(^|\()id=$id" $file || set_prop id $id $file
   fi
 done
 
 
-# update README ###
+# update README
 
-if [ README.md -ot acc/default-config.txt ] \
-  || [ README.md -ot acc/strings.sh ]
+if [ README.md -ot $id/default-config.txt ] \
+  || [ README.md -ot $id/strings.sh ]
 then
 # default config
   set -e
-  { sed -n '1,/#DC#/p' README.md; echo; cat acc/default-config.txt; \
+  { sed -n '1,/#DC#/p' README.md; echo; cat $id/default-config.txt; \
     echo; sed -n '/^#\/DC#/,$p' README.md; } > README.md.tmp
 # terminal commands
   { sed -n '1,/#TC#/p' README.md.tmp; \
-    echo; sed -n '/^Usage/,/^Run a/p' acc/strings.sh; \
+    echo; sed -n '/^Usage/,/^Run a/p' $id/strings.sh; \
     echo; sed -n '/^#\/TC#/,$p' README.md.tmp; } > README.md
     rm README.md.tmp
   set +e
 fi
 
 
-# update busybox config in scripts outside $id/
+# update busybox config (from acc/setup-busybox.sh) in $id/uninstall.sh and install scripts
 set -e
-for file in ./install-*.sh \
-  ./uninstaller-src/META-INF/com/google/android/update-binary
-do
-  if [ $file -ot acc/setup-busybox.sh ]; then
+for file in ./$id/uninstall.sh ./install-*.sh; do
+  if [ $file -ot $id/setup-busybox.sh ]; then
     { sed -n '1,/#BB#/p' $file; \
-    sed -n '/^if /,/^fi/p' acc/setup-busybox.sh; \
+    sed -n '/^if /,/^fi/p' $id/setup-busybox.sh; \
     sed -n '/^#\/BB#/,$p' $file; } > ${file}.tmp
     mv -f ${file}.tmp $file
   fi
@@ -78,25 +77,31 @@ done
 set +e
 
 
-# unify installers
+# unify installers for flashable zip (customize.sh and update-binary are copies of install-current.sh)
 { cp -u install-current.sh customize.sh
 cp -u install-current.sh META-INF/com/google/android/update-binary; } 2>/dev/null
 
 
-# initial cleanup
-rm -rf _builds/${id}-${versionCode}/ 2>/dev/null
-mkdir -p bin _builds/${id}-$versionCode/${id}-${versionCode}
-
-
-# generate $id uninstaller flashable zip
-echo "=> bin/${id}-uninstaller.zip"
-rm bin/${id}-uninstaller.zip 2>/dev/null
-(cd uninstaller-src; zip -r9 ../bin/${id}-uninstaller.zip META-INF \
-  | sed 's|^.*adding: ||' | grep -iv 'zip warning:')
-echo
+if [ bin/${id}-uninstaller.zip -ot $id/uninstall.sh ] || [ ! -f bin/${id}-uninstaller.zip ]; then
+  # generate $id uninstaller flashable zip
+  echo "=> bin/${id}-uninstaller.zip"
+  rm -rf bin/${id}-uninstaller.zip $tmpDir 2>/dev/null
+  mkdir -p bin $tmpDir
+  cp $id/uninstall.sh $tmpDir/update-binary
+  echo "#MAGISK" > $tmpDir/updater-script
+  (cd .tmp
+  zip -r9 ../bin/${id}-uninstaller.zip * \
+    | sed 's|^.*adding: ||' | grep -iv 'zip warning:')
+  rm -rf .tmp
+  echo
+fi
 
 
 if [ -z "$1" ]; then
+
+  # cleanup
+  rm -rf _builds/${id}-$versionCode/ 2>/dev/null
+  mkdir -p _builds/${id}-$versionCode/${id}-$versionCode
 
   # generate $id flashable zip
   echo "=> _builds/${id}-${versionCode}/${id}-${versionCode}.zip"
